+++ scalar/binary.go
@@ -3,7 +3,7 @@ package scalar
 import (
 	"bytes"
 
-	"github.com/apache/arrow/go/v13/arrow"
+	"github.com/apache/arrow/go/v14/arrow"
 )
 
 type Binary struct {