+++ scalar/float.go
@@ -4,7 +4,7 @@ import (
 	"math"
 	"strconv"
 
-	"github.com/apache/arrow/go/v13/arrow"
+	"github.com/apache/arrow/go/v14/arrow"
 )
 
 type Float struct {