+++ schema/errors.go
@@ -4,7 +4,7 @@ package schema
 import (
 	"fmt"
 
-	"github.com/apache/arrow/go/v13/arrow"
+	"github.com/apache/arrow/go/v14/arrow"
 )
 
 const (